Loader GUI: remove console, adjust UI to fit
